[Philip] passing data to ShowBillDetailsActivity

+++ app/src/main/java/com/example/c0778584_w2020_mad3125_fp/model/Internet.java
@@ -2,9 +2,9 @@ package com.example.c0778584_w2020_mad3125_fp.model;
 
 public class Internet extends Bill {
     private String providerName;
-    private Integer gbUsed;
+    private Double gbUsed;
 
-    public Internet(String billId, String billDate, BillType billType, String providerName, Integer gbUsed)
+    public Internet(String billId, String billDate, BillType billType, String providerName, Double gbUsed)
     {
         super(billId, billDate, billType);
         this.providerName = providerName;
@@ -13,8 +13,7 @@ public class Internet extends Bill {
     }
 
     @Override
-    public Double billCalculate()
-    {
+    public Double billCalculate(){
         double billAmount = 0.0;
         if( gbUsed < 10){
             billAmount = 3 * gbUsed;
@@ -25,23 +24,19 @@ public class Internet extends Bill {
         return billAmount;
     }
 
-    public String getProviderName()
-    {
+    public String getProviderName() {
         return providerName;
     }
 
-    public void setProviderName(String providerName)
-    {
+    public void setProviderName(String providerName) {
         this.providerName = providerName;
     }
 
-    public Integer getGbUsed()
-    {
+    public Double getGbUsed() {
         return gbUsed;
     }
 
-    public void setGbUsed(Integer gbUsed)
-    {
+    public void setGbUsed(Double gbUsed) {
         this.gbUsed = gbUsed;
     }
 }
